refactor(app): change package name to com.anangkur.movieku.consumer
