+++ app/build.gradle
@@ -6,7 +6,7 @@ apply plugin: 'kotlin-kapt'
 android {
     compileSdkVersion 29
     defaultConfig {
-        applicationId "com.anangkur.madesubmission2"
+        applicationId "com.anangkur.movieku.consumer"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
