--- app/src/main/java/com/anangkur/madesubmission2/data/local/SharedPreferenceHelper.kt
@@ -1,21 +1,0 @@
-package com.anangkur.madesubmission2.data.local
-
-import android.content.Context
-
-class SharedPreferenceHelper(context: Context){
-
-    private val prefName = "com.anangkur.madesubmission"
-    private val sharedPreferences = context.getSharedPreferences(prefName, Context.MODE_PRIVATE)
-
-    fun saveStringPreferences(key: String, language: String){
-        sharedPreferences.edit().putString(key, language).apply()
-    }
-
-    fun loadStringPreferences(key: String): String?{
-        return sharedPreferences.getString(key, "")
-    }
-
-    fun deleteStringPreferences(key: String){
-        sharedPreferences.edit().remove(key).apply()
-    }
-}
