--- app/src/main/java/com/anangkur/madesubmission2/data/model/Dates.kt
@@ -1,6 +1,0 @@
-package com.anangkur.madesubmission2.data.model
-
-data class Dates(
-    val maximum: String,
-    val minimum: String
-)
