--- app/src/main/java/com/anangkur/madesubmission2/feature/main/MainItemListener.kt
@@ -1,7 +1,0 @@
-package com.anangkur.madesubmission2.feature.main
-
-import com.anangkur.madesubmission2.data.model.Result
-
-interface MainItemListener {
-    fun onClickItem(data: Result)
-}
