+++ app/src/main/java/com/anangkur/movieku/consumer/data/model/Dates.kt
@@ -1,0 +1,6 @@
+package com.anangkur.movieku.consumer.data.model
+
+data class Dates(
+    val maximum: String,
+    val minimum: String
+)
