+++ app/src/main/java/com/anangkur/movieku/consumer/data/model/Response.kt
@@ -1,0 +1,12 @@
+package com.anangkur.movieku.consumer.data.model
+
+data class Response(
+    val success: Boolean,
+    val dates: Dates,
+    val page: Int,
+    val results: List<Result>,
+    val total_pages: Int,
+    val total_results: Int,
+    var status_message: String,
+    val status_code: Int
+)
