+++ app/src/main/java/com/anangkur/movieku/consumer/feature/detail/DetailActionListener.kt
@@ -1,0 +1,8 @@
+package com.anangkur.movieku.consumer.feature.detail
+
+import com.anangkur.movieku.consumer.data.model.Result
+
+interface DetailActionListener {
+    fun onAddFavourite(data: Result)
+    fun onRemoveFavourite(data: Result)
+}
