+++ app/src/main/java/com/anangkur/movieku/consumer/feature/main/MainItemListener.kt
@@ -1,0 +1,7 @@
+package com.anangkur.movieku.consumer.feature.main
+
+import com.anangkur.movieku.consumer.data.model.Result
+
+interface MainItemListener {
+    fun onClickItem(data: Result)
+}
